Correção de bugs

+++ app/src/main/java/pt/ipt/DAMA/model/astronomyAPI/AstronomyPositionResponseDTO.kt
@@ -21,7 +21,7 @@ data class AstronomyPositionResponseDTO(
             data class Location(
                 @SerializedName("longitude") val longitude: Double,
                 @SerializedName("latitude") val latitude: Double,
-                @SerializedName("elevation") val elevation: Int
+                @SerializedName("elevation") val elevation: Double
             )
         }
